requirements.txt
requests==2.32.3
python-dotenv==1.0.1
selenium==4.27.1
selenium-wire==5.1.0
Brotli==1.1.0
blinker==1.7.0